+++ src/core/types/route.ts
@@ -1,0 +1,8 @@
+type Route = {
+    path: string;
+    element: HTMLElement;
+    default?: boolean
+
+}
+
+export default Route;
